# Please enter the commit message for your changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit. # Explicit paths specified without -i or -o; assuming --only paths... # On branch issue-13153-read-book-on-ground # Changes to be committed: #	modified:   CataclysmWin.cbp -- to be ignored due to personal settings #	modified:   src/game.cpp     -- change whole function game::read() #	modified:   src/game.h       -- add friend class player_activity; // To allow drop of items in player_activity::finish #	modified:   src/player.h     -- add public variable bool reading_from_inventory= false; #	modified:   src/player_activity.cpp  -- add to player_activity::finish under ACT_READ # #                // below code added in to support reading from ground/vehicles #                if (!p->reading_from_inventory ){ #                    g->drop(p->inv.size()-1); #                } #                else { #                    p->reading_from_inventory = false; #                } # #

+++ src/game.cpp
@@ -11639,7 +11639,7 @@ void game::read()
             vehicle_stack vs = veh->get_items( part );
 
             // check if item is in vehicle
-            for (item_index = 0; item_index < vs.size();item_index++) {
+            for (item_index = 0; item_index < (int) vs.size();item_index++) {
                 if (it == &vs[item_index]){
                     item_on_veh = true;
                     break;
@@ -11653,7 +11653,7 @@ void game::read()
         map_stack ms = m.i_at(u.pos() );
 
         // check if item is on ground
-        for (item_index = 0; item_index < ms.size();item_index++) {
+        for (item_index = 0; item_index < (int) ms.size();item_index++) {
                 if (it == &ms[item_index]){
                     break;
                 }
